+++ apps/web/src/styles/dist.css
@@ -1154,6 +1154,10 @@ video {
   display: contents;
 }
 
+.hidden {
+  display: none;
+}
+
 .h-1\/2 {
   height: 50%;
 }
