+++ packages/react-opensheet-music-display/src/machines/cursor.machine.typegen.ts
@@ -36,7 +36,7 @@
         eventsCausingServices: {
           
         };
-        matchesStates: "idle";
+        matchesStates: "idle" | "initializing";
         tags: never;
       }
