+++ packages/ui/components/Exercise/index.tsx
@@ -1,1 +1,3 @@
+'use client';
+
 export * from './Exercise';
